- implemented rendering of profile data, received from backend, at frontend; - when user adds a new Education, it is saved to Redux and rendered on Ui straight away; - when user clicks on button to view his own profile, he is redirected to /profiles/usedId page (page with his profile, based on his id)
