+++ src/frontend/src/redux/Profile/profileActionTypes.js
@@ -1,2 +1,3 @@
 export const SAVE_ACTIVE_PROFILE = 'SAVE_ACTIVE_PROFILE'
 export const LOADING_PROFILE = 'LOADING_PROFILE'
+export const ADD_NEW_EDUCATION = 'ADD_NEW_EDUCATION'
