+++ src/frontend/src/redux/Profile/profileActions.js
@@ -1,6 +1,7 @@
 import * as actions from './profileActionTypes'
 import http from '../../services/httpService'
 import getHeaders from '../../services/headersService'
+import convertStringsToLocalDate from '../../utils/convertStringsToLocalDate'
 
 export const getActiveProfileAction = (userId) => (dispatch) => {
   dispatch({type: actions.LOADING_PROFILE, payload: true})
@@ -12,5 +13,18 @@ export const getActiveProfileAction = (userId) => (dispatch) => {
         payload: res.data
       })
       dispatch({ type: actions.LOADING_PROFILE, payload: false })
+    })
+}
+
+export const createNewEducationAction = (payload) => (dispatch) => {
+  convertStringsToLocalDate(payload)
+
+  return http
+    .post('/api/educations', payload,
+      {headers: getHeaders()}
+    )
+    .then((res) => res.data)
+    .then((education) => {
+      dispatch({ type: actions.ADD_NEW_EDUCATION, payload: education })
     })
 }
