+++ src/frontend/src/redux/Profile/profileReducer.js
@@ -1,7 +1,7 @@
 import * as actions from './profileActionTypes'
 
 const initialState = {
-  loading: false,
+  loading: true,
   activeProfile: {} // extended userRs object with many nested fields like educations, certifications, etc.
 }
 
@@ -15,6 +15,15 @@ const profileReducer = (state = initialState, action) => {
 
     case actions.LOADING_PROFILE:
       return { ...state, loading: action.payload }
+
+    case actions.ADD_NEW_EDUCATION:
+      return {
+        ...state,
+        activeProfile: {
+          ...state.activeProfile,
+          educations: [...state.activeProfile.educations, action.payload]
+        }
+      }
 
     default: {
       return state
