+++ src/frontend/src/redux/User/userActionTypes.js
@@ -1,5 +1,3 @@
 export const SAVE_ACTIVE_USER = 'SAVE_ACTIVE_USER'
 export const SIGN_OUT = 'SIGN_OUT'
 export const LOADING_USERS = 'LOADING_USERS'
-
-export const ADD_NEW_EDUCATION = 'ADD_NEW_EDUCATION'
