+++ src/frontend/src/redux/User/userActions.js
@@ -1,7 +1,6 @@
 import * as actions from './userActionTypes'
 import http from '../../services/httpService'
 import getHeaders from '../../services/headersService'
-import convertStringsToLocalDate from '../../utils/convertStringsToLocalDate'
 
 export const getActiveUserAction = () => (dispatch) => {
   dispatch({type: actions.LOADING_USERS, payload: true})
@@ -21,17 +20,4 @@ export const signOutAction = () => (dispatch) => {
     type: actions.SIGN_OUT,
     payload: null
   })
-}
-
-export const createNewEducationAction = (payload) => (dispatch) => {
-  convertStringsToLocalDate(payload)
-
-  return http
-    .post('/api/educations', payload,
-      {headers: getHeaders()}
-    )
-    .then((res) => res.data)
-    .then((education) => {
-      dispatch({ type: actions.ADD_NEW_EDUCATION, payload: education })
-    })
 }
