+++ src/frontend/src/redux/User/userReducer.js
@@ -2,29 +2,7 @@ import * as actions from './userActionTypes'
 
 const initialState = {
   loadingUser: false,
-  activeUser: {},
-  educations: [
-    { id: 7,
-      school: 'Boston University',
-      degreeReceived: 'Master`s',
-      activities: 'Football',
-      description: 'sdf',
-      fieldOfStudy: 'Business',
-      dateStart: '2001',
-      dateFinish: '2004'
-    },
-    {
-      school: 'DANIT',
-      degreeReceived: 'Master`s',
-      activities: 'Football',
-      description: 'sdf',
-      fieldOfStudy: 'Business',
-      dateStart: '2019',
-      dateFinish: '2021'
-    }
-  ],
-  certifications: [],
-  workPlaces: []
+  activeUser: {}
 }
 
 const userReducer = (state = initialState, action) => {
@@ -42,14 +20,6 @@ const userReducer = (state = initialState, action) => {
         activeUser: action.payload
       }
 
-    case actions.ADD_NEW_EDUCATION:
-      return {
-        ...state,
-        educations:
-              [ state.educations
-                ? [ ...state.educations, action.payload ]
-                : action.payload]
-      }
     default: {
       return state
     }
